Persist avatar
Comment in actor-document create method to set a default. Unlike in
forbidden lands, which simply uses `img` in their template, we needed to
use `this.actor.img`, presumably because our template is in a different
scope. Also, switched to the homer image because the original errors for
not having a file extension.