Add bonds
Treated as a type of item, much like armor, spell, or class, bonds link
characters in up to three ways to other entities. The source material
does not give a name to each of the three axes that make up a bond, so
I've chosen a name for each as follows. Let me know what you think!

```
Admiration vs. Inferiority => Respect
Loyalty vs. Mistrust => Trust
Affection vs. Hatred => Liking
```

On each axis, the value is stored as a string - either an empty string
to represent no preference, or the positive or negative affect.