Adds the cheat sheet tab
Moving some info into (unstyled) tooltips, this adds the rest into its
own tab. The dedicated tab only lists available actions as well as
opportunities, but it includes everything from the reference doc with
the exception of checks. I feel like that's mostly useful for a DM, so
I'm not sure whether it should go on there, but I have no problem adding
it. To also add is a custom area for notes maybe?

Of note, removing the custom css causes the pronouns section not to run
up into the tabs. Some formatting still needs to be done however.