Replace grid-template-areas with grid-column
Open to opinions, but personally I found the grid-template-areas really
confusing, which isn't a reason not to use them, but I do think it's
much simpler to simply allow the fields to be displayed in the order
they're defined and mark the ones that are multiple columns.

This unfortunately doesn't work with attributes_and_status_effects and
the footer because there are blank spaces. Maybe there's another
solution for those that isn't defining an empty div?

What do y'all think? And is there a better name than
double-/triple-wide?